+++ util-bulk.R
@@ -15,6 +15,7 @@
 ## Copyright 2017 by Raphael Nömmer <noemmer@fim.uni-passau.de>
 ## Copyright 2017 by Christian Hechtl <hechtl@fim.uni-passau.de>
 ## Copyright 2019 by Thomas Bock <bockthom@fim.uni-passau.de>
+## Copyright 2021 by Niklas Schneider <s8nlschn@stud.uni-saarland.de>
 ## All Rights Reserved.
 
 
@@ -45,7 +46,7 @@ collect.multi.networks = function(project.conf, network.conf, step = 1) {
     ## collect the network objects for all the ranges
     networks = lapply(ranges, function(range) {
         ## construct range data
-        range.data = RangeData$new(project.conf, range)
+        range.data = RangeData$new(project.conf, range, built.from.range.data.read = TRUE)
 
         ## construct network builder
         network.builder <- NetworkBuilder$new(range.data, network.conf)
@@ -88,7 +89,7 @@ collect.bipartite.networks = function(project.conf, network.conf, step = 1) {
     ## collect the network objects for all the ranges
     networks = lapply(ranges, function(range) {
         ## construct range data
-        range.data = RangeData$new(project.conf, range)
+        range.data = RangeData$new(project.conf, range, built.from.range.data.read = TRUE)
 
         ## construct network builder
         network.builder <- NetworkBuilder$new(range.data, network.conf)
@@ -131,7 +132,7 @@ collect.author.networks = function(project.conf, network.conf, step = 1) {
     ## collect the network objects for all the ranges
     networks = lapply(ranges, function(range) {
         ## construct range data
-        range.data = RangeData$new(project.conf, range)
+        range.data = RangeData$new(project.conf, range, built.from.range.data.read = TRUE)
 
         ## construct network builder
         network.builder <- NetworkBuilder$new(range.data, network.conf)
@@ -174,7 +175,7 @@ collect.artifact.networks = function(project.conf, network.conf, step = 1) {
     ## collect the network objects for all the ranges
     networks = lapply(ranges, function(range) {
         ## construct range data
-        range.data = RangeData$new(project.conf, range)
+        range.data = RangeData$new(project.conf, range, built.from.range.data.read = TRUE)
 
         ## construct network builder
         network.builder <- NetworkBuilder$new(range.data, network.conf)
@@ -221,7 +222,7 @@ construct.data = function(project.conf, callgraphs = FALSE, step = 1) {
                                     "")
 
         ## construct range data
-        range.data = RangeData$new(project.conf, range, revision.callgraph)
+        range.data = RangeData$new(project.conf, range, revision.callgraph, built.from.range.data.read = TRUE)
         attr(range.data, "range") = range
 
         # add to global list